requirements.txt
google-generativeai==0.3.2
gspread==5.12.0
oauth2client==4.1.3
Pillow==10.1.0
requests==2.31.0